# src/main/resources/api.yaml
...
swagger: "2.0"
info:
  description: First API for the Tracey database models.
  version: 1.0.0
  title: Tracey_Database_API
  contact:
    email: fish_b@it.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /rikkygaray21/tracey-01/1.0.0
...
schemes:
- https
paths:
...
  /users/{userId}/days/{dayId}/activities:
    get:
      summary: Get all activites for a specific user on a given day searching by user and day Id.
      operationId: getActivitiesByUserAndDay
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      responses:
        "200":
          description: ActivitiesByUserAndDay - Get activities on specific day - success
        "400":
          description: ActivitiesByUserAndDay - Get activities on specific day - failed
    post:
      summary: Create and add an activity for an user on a given day.
      operationId: postActivitiyByUserAndDay
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      responses:
        "200":
          description: ActivitiesByUserAndDay - Post activities on specific day - success
        "400":
          description: ActivitiesByUserAndDay - Post activities on specific day - failed
  /users/{userId}/months/{monthId}/activities:
    get:
      summary: Gets all activites for a specific user on a given month by searching its user and month Id.
      operationId: getActivitiesByUserAndMonth
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      responses:
        "200":
          description: ActivitiesByUserAndMonth - Get activities on specific month - success
        "400":
          description: ActivitiesByUserAndMonth - Get activities on specific month - failed
    post:
      summary: Create and add an activity for a specific user on a given month by searching its user and month Id.
      operationId: postActivitiesByUserAndMonth
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      responses:
        "200":
          description: ActivitiesByMonth - Post specific users activities on specific month - success
        "400":
          description: ActivitiesByMonth - Post specific users activities on specific month - failed
  /users/{userId}/days/{dayId}/activites/{activityId}:
    get:
      summary: Get a specific activity. Find by User, Day and Activity Id.
      operationId: getActivityByIdAndUserAndDayId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndUserAndDayId - Get specific users activity on specific day - success
        "400":
          description: ActivityByIdAndUserAndDayId - Get specific users activity on specific day - failed
    put:
      summary: Modify/Put a specific activity. Find by User, Day and Activity Id.
      operationId: putActivityByIdAndUserAndDayId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndUserAndDayId - Put specific users activity on specific day - success
        "400":
          description: ActivityByIdAndUserAndDayId - Put specific users activity on specific day - failed
    delete:
      summary: Delete a specific activity by its Id
      operationId: deleteActivityByIdAndByDay
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndDayId - Delete specific users activity on specific day - success
        "400":
          description: ActivityByIdAndDayId - Delete specific users activity on specific day - failed
    patch:
      summary: Modify/Patch a specific activity. Find by User, Day and Activity Id.
      operationId: patchActivityByIdAndUserAndDayId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: dayId
        in: path
        description: String ID of the day to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndUserAndDayId - Patch specific users activity on specific day - success
        "400":
          description: ActivityByIdAndUserAndDayId - Patch specific users activity on specific day - failed
  /users/{userId}/months/{monthId}/activites/{activityId}:
    get:
      summary: Get a specific activtiy. Find by searching for User, Month and Activity Id.
      operationId: getActivityByIdAndUserAndMonthId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndMonthAndDay - Get specific users activity on specific month - success
        "400":
          description: ActivityByIdAndDayId - Get specific users activity on specific month - failed
    put:
      summary: Modify/Put a specific activtiy. Find by searching for User, Month and Activity Id.
      operationId: putActivityByIdAndUserAndMonthId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndDayId - Put specific users activity on specific month - success
        "400":
          description: ActivityByIdAndDayId - Put specific users activity on specific month - failed
    delete:
      summary: Delete a specific activtiy. Find by searching for User, Month and Activity Id.
      operationId: deleteActivityByIdAndUserAndMonthId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndDayId - Delete specific activity on specific month - success
        "400":
          description: ActivityByIdAndDayId - Delete specific activity on specific month - failed
    patch:
      summary: Modify/Patch a specific activtiy. Find by searching for User, Month and Activity Id.
      operationId: patchActivityByIdAndUserAndMonthId
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      - name: monthId
        in: path
        description: String ID of the month to get
        required: true
        type: string
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivityByIdAndDayId - Patch specific activity on specific month - success
        "400":
          description: ActivityByIdAndDayId - Patch specific activity on specific month - failed
  /users:
    get:
      summary: Get all user records
      operationId: getUsersRecord
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Users - Get users record - success
        "400":
          description: Users - Get users record - failed
    post:
      summary: Create a user record
      operationId: postUsersRecord
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Users - Post users record - success
        "400":
          description: Users - Post users record - failed
  /users/{userId}:
    get:
      summary: Gets a specific user by its Id
      operationId: getUserById
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to get
        required: true
        type: string
      responses:
        "200":
          description: UsersById - Get users record - success
        "400":
          description: UsersById - Get users record - failed
    put:
      summary: Modify/Put a specific user by its Id
      operationId: putUserById
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to put
        required: true
        type: string
      responses:
        "200":
          description: UsersById - Put users record - success
        "400":
          description: UsersById - Put users record - failed
    delete:
      summary: Delete a specific user by its Id
      operationId: deleteUserById
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to delete
        required: true
        type: string
      responses:
        "200":
          description: UsersById - Delete users record - success
        "400":
          description: UsersById - Delete users record - failed
    patch:
      summary: Modify/Patch a specific user by its Id
      operationId: patchUserById
      produces:
      - application/json
      parameters:
      - name: userId
        in: path
        description: String ID of the user to patch
        required: true
        type: string
      responses:
        "200":
          description: UsersById - Patch users record - success
        "400":
          description: UsersById - Patch users record - failed
...